requirements/base.txt
django==4.2.9
djangorestframework==3.15.2
django-crispy-forms==2.3
django-environ==0.9.0
django-filter==24.3
python-dotenv==0.21.0
psycopg==3.2.1
selenium==4.23.1
